refactor(grids): move s3 methods to shared class
